nvchecker: port over to v2 nvchecker
Which for inexplicable reasons decided to inflict the ugliness of *.toml
on me. So yay for pointlessly reformatting every line in the file, and
making it twice as verbose and half as readable.

This would be livable if there were any rationale offered for why the
switch was made, but @lilydjwg has merely noted in the documentation "it
has been changed because I said so, here's a script to migrate your
config file and totally fail to produce something usable".

Various settings or entire sources have thus been commented out because
they don't work, until I figure out how to make them work or even if
they work.

consolation prize: there is now a default config file